// src/shared/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        Module: {
            id?: number;
            title?: string;
            slug?: string;
            order?: number;
            /** @enum {string} */
            subject?: "math_literacy" | "profile_math";
            lesson_count?: number;
        };
        ModuleDetail: components["schemas"]["Module"] & {
            lessons?: components["schemas"]["LessonSummary"][];
        };
// ...
    };
    responses: {
        /** @description Error */
        Error: {
            headers: {
                [name: string]: unknown;
            };
            content: {
                "application/json": components["schemas"]["Error"];
            };
        };
    };
    parameters: {
        IdPath: number;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}